Blogger can create new posts.
Controller returns a 201 status object containing a json object with a success property and the post after successfully creating the post with mongoose.

On failure, controller returns a 400 status object containing a json object with a success property and array of error messages.

NewPostPage navigates to the PostPage of the new post on success or displays errors in the NewPostPage.

Removed post_create_get from blogpostController and blog.js routes

Post page and home page now display the timestamp of the post.